Make sub-expressions callable
This PR extends the Handlebars language to allow sub-expressions to be
callable values. Example:

```handlebars
Nesting one layer deep:

{{(helper)}}

Nesting two layers deep:

{{((helper))}}

Nesting as an argument value:

{{helper arg=((sub-helper foo) bar)}}
```

This is to enable this capability in Ember/Glimmer templates, since
there are cases where it will be necessary for us to implement template
strict mode.