Fix whitespace control for chained blocks
omitLeft failed for helper ending block in chained inverse. It tried to omitLeft whitespace only for the first inverse, which is incorrect.
Failing scenario:
```
{{#if false}}
{{else if false}}
{{else}}
	{{/if}}
```
TAB used to make it into the template.
WhitespaceControl failed to omit tab character, while working properly on a bit simpler scenario:
```
{{#if false}}
{{else}}
	{{/if}}
```
Here, TAB did not use to make it into the template.